Add file upload functionality with status notifications and styling
