+++ www/assets/css/gallery.scss
@@ -1,4 +1,5 @@
 @use "ceci-variables.scss";
+@use "base-button.scss";
 
 .grid-container {
   columns: 5 200px;
@@ -45,94 +46,4 @@
 #pages {
   width: 100%;
   margin: 10px auto;
-}
-
-button {
-  background-color: ceci-variables.$button-bg-color;
-  color: ceci-variables.$button-color;
-  border: 2px solid ceci-variables.$button-color;
-  padding: 5px;
-  margin: 0 1px 0 1px;
-  width: auto;
-  cursor: pointer;
-  align-items: center;
-  justify-content: center;
-  border-radius: 5px;
-
-  &:hover {
-    color: ceci-variables.$button-color-hover;
-    stroke: ceci-variables.$button-color-hover;
-    background-color: ceci-variables.$button-bg-color-hover;
-    border: 2px solid ceci-variables.$button-border-hover;
-  }
-
-  &:active {
-    stroke: white;
-    color: white;
-    background-color: ceci-variables.$button-color;
-    border: 2px solid ceci-variables.$button-border-active;
-  }
-}
-
-button.selected {
-  @extend button;
-  background-color: ceci-variables.$button-color;
-  color: ceci-variables.$button-bg-color;
-
-  &:hover {
-    background-color: ceci-variables.$button-bg-color-hover;
-    color: ceci-variables.$button-color-hover;
-    stroke: ceci-variables.$button-color-hover;
-    border: 2px solid ceci-variables.$button-border-hover;
-  }
-}
-
-.file-upload-btn {
-  @extend button;
-  stroke: ceci-variables.$button-color;
-  display: inline-flex;
-}
-
-.status {
-  position: fixed;
-  display: none;
-  padding: 5px 10px;
-  text-align: center;
-  border-radius: 5px;
-  font-size: 14px;
-  color: #fff;
-  left: 0;
-  right: 0;
-  top: 0;
-  background-color: #333;
-  z-index: 1000;
-}
-
-.success {
-  background-color: ceci-variables.$success;
-  /* Green */
-}
-
-.error {
-  background-color: ceci-variables.$error;
-  /* Yellow */
-  color: ceci-variables.$error-color;
-  /* Black text color for better contrast */
-}
-
-#imageInput,
-#imageUpload {
-  display: none;
-  /* Hide the file input */
-}
-
-.visually-hidden {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  margin: -1px;
-  padding: 0;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  border: 0;
 }
