+++ www/content/es/fotos.md
@@ -4,4 +4,5 @@ title: "Fotos"
 
 Envíanos tus fotos de la fiesta a través de nuestro <a href="https://t.me/cecinestpasun_bot" target="_blank">bot de Telegram</a> y aparecerán aquí automáticamente.
 
+{{< upload >}}
 {{< gallery >}}
