+++ www/content/es/fotoup.md
@@ -1,0 +1,5 @@
+---
+title: "Foto upload"
+---
+
+{{< upload 54 >}}
