+++ www/layouts/shortcodes/upload.html
@@ -1,0 +1,101 @@
+{{ $size := .Get 0 | default 24 }} {{ $opts := dict "transpiler" "dartsass" }}{{
+with resources.Get "css/upload.scss" | css.Sass $opts | resources.Minify |
+resources.Fingerprint }}
+<link
+  rel="stylesheet"
+  href="{{ .RelPermalink }}"
+  integrity="{{ .Data.Integrity }}"
+/>
+{{ end }}
+
+<svg
+  title="Send a picture"
+  tabindex="0"
+  role="button"
+  id="cameraUpload"
+  class="file-upload-btn"
+  aria-controls="imageInput"
+  aria-label="Upload image from camera"
+  xmlns="http://www.w3.org/2000/svg"
+  viewBox="0 0 24 24"
+  width="{{ $size }}"
+  height="{{ $size }}"
+  fill="none"
+>
+  <path
+    d="M12 16C13.6569 16 15 14.6569 15 13C15 11.3431 13.6569 10 12 10C10.3431 10 9 11.3431 9 13C9 14.6569 10.3431 16 12 16Z"
+    vector-effect="non-scaling-stroke"
+    stroke-width="2"
+    stroke-linecap="round"
+    stroke-linejoin="round"
+  />
+  <path
+    d="M3 16.8V9.2C3 8.0799 3 7.51984 3.21799 7.09202C3.40973 6.71569 3.71569 6.40973 4.09202 6.21799C4.51984 6 5.0799 6 6.2 6H7.25464C7.37758 6 7.43905 6 7.49576 5.9935C7.79166 5.95961 8.05705 5.79559 8.21969 5.54609C8.25086 5.49827 8.27836 5.44328 8.33333 5.33333C8.44329 5.11342 8.49827 5.00346 8.56062 4.90782C8.8859 4.40882 9.41668 4.08078 10.0085 4.01299C10.1219 4 10.2448 4 10.4907 4H13.5093C13.7552 4 13.8781 4 13.9915 4.01299C14.5833 4.08078 15.1141 4.40882 15.4394 4.90782C15.5017 5.00345 15.5567 5.11345 15.6667 5.33333C15.7216 5.44329 15.7491 5.49827 15.7803 5.54609C15.943 5.79559 16.2083 5.95961 16.5042 5.9935C16.561 6 16.6224 6 16.7454 6H17.8C18.9201 6 19.4802 6 19.908 6.21799C20.2843 6.40973 20.5903 6.71569 20.782 7.09202C21 7.51984 21 8.0799 21 9.2V16.8C21 17.9201 21 18.4802 20.782 18.908C20.5903 19.2843 20.2843 19.5903 19.908 19.782C19.4802 20 18.9201 20 17.8 20H6.2C5.0799 20 4.51984 20 4.09202 19.782C3.71569 19.5903 3.40973 19.2843 3.21799 18.908C3 18.4802 3 17.9201 3 16.8Z"
+    vector-effect="non-scaling-stroke"
+    stroke-width="2"
+    stroke-linecap="round"
+    stroke-linejoin="round"
+  />
+</svg>
+<label for="imageInput" class="visually-hidden">Select image to upload</label>
+<input
+  type="file"
+  id="imageInput"
+  name="image"
+  accept="image/*"
+  capture
+  text=""
+  aria-hidden="true"
+  tabindex="-1"
+/>
+<!-- this one works with camera for chrome, but it (silently) fails for firefox front camera -->
+
+<svg
+  title="Send a picture from device"
+  tabindex="0"
+  role="button"
+  id="customPictureUpload"
+  class="file-upload-btn"
+  araia-controls="imageInput"
+  aria-label="Upload image from file"
+  width="{{ $size }}"
+  height="{{ $size }}"
+  viewBox="0 0 24 24"
+  fill="none"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    d="M13.6136 15.3473L17.8651 10.9766L21 13.9844M6.96484 19L11.9688 13.9766L17.9727 19M9.96875 9.97656C9.96875 11.0811 9.07332 11.9766 7.96875 11.9766C6.86418 11.9766 5.96875 11.0811 5.96875 9.97656C5.96875 8.87199 6.86418 7.97656 7.96875 7.97656C9.07332 7.97656 9.96875 8.87199 9.96875 9.97656ZM12.0627 6.06274L11.9373 5.93726C11.5914 5.59135 11.4184 5.4184 11.2166 5.29472C11.0376 5.18506 10.8425 5.10425 10.6385 5.05526C10.4083 5 10.1637 5 9.67452 5H6.2C5.0799 5 4.51984 5 4.09202 5.21799C3.71569 5.40973 3.40973 5.71569 3.21799 6.09202C3 6.51984 3 7.07989 3 8.2V15.8C3 16.9201 3 17.4802 3.21799 17.908C3.40973 18.2843 3.71569 18.5903 4.09202 18.782C4.51984 19 5.07989 19 6.2 19H17.8C18.9201 19 19.4802 19 19.908 18.782C20.2843 18.5903 20.5903 18.2843 20.782 17.908C21 17.4802 21 16.9201 21 15.8V10.2C21 9.0799 21 8.51984 20.782 8.09202C20.5903 7.71569 20.2843 7.40973 19.908 7.21799C19.4802 7 18.9201 7 17.8 7H14.3255C13.8363 7 13.5917 7 13.3615 6.94474C13.1575 6.89575 12.9624 6.81494 12.7834 6.70528C12.5816 6.5816 12.4086 6.40865 12.0627 6.06274Z"
+    stroke-width="2"
+    stroke-linecap="round"
+    vector-effect="non-scaling-stroke"
+    stroke-linejoin="round"
+  />
+</svg>
+<label for="imageUpload" class="visually-hidden">Select image to upload</label>
+<input
+  type="file"
+  id="imageUpload"
+  multiple="multiple"
+  name="image"
+  accept="image/*"
+  text=""
+  aria-hidden="true"
+  tabindex="-1"
+/>
+<!-- This one works in firefox and allows rear/front camera + open from file, it is needed in chrome to allow uploading a file -->
+
+{{ $dev := (eq hugo.Environment "development") }} {{/* minify only in production
+*/}} {{ $opts := dict "minify" (not $dev) }} {{ $galleryScript := resources.Get
+"js/upload.ts" }} {{/* {{ $imgUpload := resources.Get "js/imgUpload.ts" }} */}}
+{{ $name := cond $dev "upload.ts" "upload.min.ts" }} {{/* concat the Typescript
+so that it does not repeat imports */}} {{ $js := slice $galleryScript |
+resources.Concat (printf "js/%s" $name) | js.Build $opts }} {{/* add Fingerprint
+only in production */}} {{ $js := cond $dev $js (resources.Fingerprint $js) }}
+
+<script
+  src="{{ $js.RelPermalink }}"
+  integrity="{{ $js.Data.Integrity }}"
+  crossorigin="anonymous"
+  defer
+></script>
